Edit Phones Update
I have added in the EditItemServlet and the edit-item.jsp. I had some issues with the LocalDate. Because the class is passed from NavigationServlet it does not have the year, day, month attributes so the user must edit the LocalDate from SQL. My solution was to get the releaseDate as a string and substr through it to find the date and pass those into the LocalDate.